refactor: fix mypy errors and format
