+++ moccasin/msig_cli/tx/tx_build.py
@@ -1,23 +1,23 @@
 from typing import Optional
 
 from eth_typing import ChecksumAddress
-
+from eth_utils import to_checksum_address
+from hexbytes import HexBytes
 from prompt_toolkit import HTML, PromptSession, print_formatted_text
 from prompt_toolkit.shortcuts import clear as prompt_clear
-
 from safe_eth.safe import Safe, SafeTx
 from safe_eth.safe.multi_send import MultiSend, MultiSendOperation
 from safe_eth.util.util import to_0x_hex_str
 
 from moccasin.logging import logger
-from moccasin.msig_cli.tx.prompts import (
-    prompt_operation_type,
-    prompt_safe_nonce,
+from moccasin.msig_cli.tx.build_prompts import (
     prompt_gas_token,
     prompt_internal_txs,
-    prompt_single_internal_tx,
-    prompt_save_eip712_json,
     prompt_multisend_batch_confirmation,
+    prompt_operation_type,
+    prompt_safe_nonce,
+    prompt_save_eip712_json,
+    prompt_single_internal_tx,
     prompt_target_contract_address,
 )
 from moccasin.msig_cli.tx.helpers import pretty_print_safe_tx
@@ -75,7 +75,7 @@ def run(
     value: Optional[int] = None,
     operation: Optional[int] = None,
     safe_nonce: Optional[int] = None,
-    data: Optional[bytes] = None,
+    data: Optional[HexBytes] = None,
     gas_token: Optional[str] = None,
     eip712_json_out: Optional[str] = None,
 ) -> SafeTx:
@@ -94,22 +94,29 @@ def run(
     :return: An instance of SafeTx.
     :raises GoBackToPrompt: If we ne to go back to the main CLI prompt.
     """
-    safe_nonce = prompt_safe_nonce(
-        prompt_session, safe_instance, safe_nonce, cmd="tx_builder"
-    )
-    gas_token = prompt_gas_token(prompt_session, gas_token, cmd="tx_builder")
+    # Get input values from args
+    tx_to = to_checksum_address(to) if to else None
+    tx_gas_token = to_checksum_address(gas_token) if gas_token else None
+    tx_value = value if value is not None else 0
+    tx_operation = operation
+    tx_data = data
+    tx_safe_nonce = safe_nonce
+
+    # Prompt for nonce and gas token
+    if tx_safe_nonce is None:
+        tx_safe_nonce = prompt_safe_nonce(prompt_session, safe_instance)
+    if not tx_gas_token:
+        tx_gas_token = prompt_gas_token(prompt_session)
 
     # If no data provided, we prompt for internal transactions
-    if not data:
-        internal_txs = prompt_internal_txs(
-            prompt_session, prompt_single_internal_tx, cmd="tx_builder"
-        )
+    if tx_data is None:
+        internal_txs = prompt_internal_txs(prompt_session, prompt_single_internal_tx)
         if len(internal_txs) > 1:
             multi_send = MultiSend(ethereum_client=safe_instance.ethereum_client)
-            data = multi_send.build_tx_data(internal_txs)
-            to = multi_send.address
-            value = 0
-            operation = int(
+            tx_data = multi_send.build_tx_data(internal_txs)
+            tx_to = multi_send.address
+            tx_operation = 0
+            tx_operation = int(
                 MultiSendOperation.CALL.value
                 if multi_send.call_only
                 else MultiSendOperation.DELEGATE_CALL.value
@@ -119,12 +126,12 @@ def run(
                     "\n<b><green>MultiSend transaction created successfully!</green></b>\n"
                 )
             )
-        elif len(internal_txs) == 1:
+        else:
             multi_send_one_tx = internal_txs[0]
-            to = multi_send_one_tx.to
-            value = multi_send_one_tx.value
-            data = multi_send_one_tx.data
-            operation = multi_send_one_tx.operation.value
+            tx_to = multi_send_one_tx.to
+            tx_value = multi_send_one_tx.value
+            tx_data = multi_send_one_tx.data
+            tx_operation = multi_send_one_tx.operation.value
             print_formatted_text(
                 HTML(
                     "\n<b><green>Single internal transaction created successfully!</green></b>\n"
@@ -132,27 +139,26 @@ def run(
             )
 
     # If data is provided, try to decode/confirm MultiSend batch
-    if data:
+    if tx_data is not None:
         to_decoded, op_decoded = decode_and_confirm_multisend_batch(
-            prompt_session, safe_instance, data, to, operation
+            prompt_session, safe_instance, tx_data, tx_to, tx_operation
         )
-        if to_decoded is not None:
-            to, operation = to_decoded, op_decoded
+        if to_decoded is not None and op_decoded is not None:
+            tx_to, tx_operation = to_decoded, op_decoded
 
     # If still missing, prompt for target contract address and/or operation
-    if not to:
-        to = prompt_target_contract_address(prompt_session)
-        to = ChecksumAddress(to)
-    if operation is None:
-        operation = prompt_operation_type(prompt_session)
+    if not tx_to:
+        tx_to = prompt_target_contract_address(prompt_session)
+    if tx_operation is None:
+        tx_operation = prompt_operation_type(prompt_session)
     try:
         safe_tx = safe_instance.build_multisig_tx(
-            to=to,
-            value=value,
-            operation=operation,
-            safe_nonce=safe_nonce,
-            data=data,
-            gas_token=gas_token,
+            to=tx_to,
+            value=tx_value,
+            operation=tx_operation,
+            safe_nonce=tx_safe_nonce,
+            data=tx_data,
+            gas_token=tx_gas_token,
         )
     except Exception as e:
         logger.error(f"Failed to create SafeTx instance: {e}")
