+++ tests/unit/test_unit_msig.py
@@ -1,27 +1,27 @@
 import pytest
-
 from prompt_toolkit import PromptSession
+
 from moccasin.msig_cli.validators import (
-    is_valid_address,
-    is_valid_rpc_url,
-    is_valid_number,
-    is_valid_not_zero_number,
-    is_valid_operation,
-    is_valid_data,
-    is_valid_transaction_type,
-    is_valid_function_signature,
-    is_valid_boolean,
     is_not_empty,
+    is_valid_address,
+    is_valid_boolean,
+    is_valid_data,
+    is_valid_function_signature,
+    is_valid_not_zero_number,
+    is_valid_number,
+    is_valid_operation,
+    is_valid_rpc_url,
+    is_valid_transaction_type,
     validator_address,
-    validator_rpc_url,
-    validator_number,
-    validator_not_zero_number,
-    validator_operation,
-    validator_data,
-    validator_transaction_type,
-    validator_function_signature,
     validator_boolean,
+    validator_data,
+    validator_function_signature,
     validator_not_empty,
+    validator_not_zero_number,
+    validator_number,
+    validator_operation,
+    validator_rpc_url,
+    validator_transaction_type,
 )
